OB-35: Move login page related files under app/components/login
