+++ app/app.scss
@@ -1,7 +1,7 @@
 @import "variables";
 @import "components/header";
 @import "components/list";
-@import "components/login";
+@import "components/login/login";
 
 body {
     margin: 0;
